move trace manager to trace package
